(PC-12840) feat(accessibility): Update web TabBar tests
